+++ app/controllers/leagues_controller.rb
@@ -639,6 +639,7 @@ class LeaguesController < ApplicationController
                                    :league_id_preround, :has_preround, :preround_point_modus, :preround_scorer_modus,
                                    :league_id_direct_encounters,
                                    :table_modus, :direct_comparison, :periods, :period_length, :overtime_length,
+                                   :game_duration_minutes,
                                    :banner_link_url, :parental_consent_required,
                                    required_documents: [])
   end